Another round of debugging. This time things are largely working, although getting there took some effort. Fixed some errors in the physics code as well as a bad render loop. There's a massive performance issue with how the load queues are managed, but dealing with that's been deferred for now by just massively restricting what gets loaded.
-Peter Mawhorter

+++ src/datatypes/octree.c
@@ -79,7 +79,7 @@ void oct_insert_recursive(octree *ot, void *object, bbox *box) {
   ot->count += 1;
 }
 
-int oct_remove_recursive(octree *ot, void *object) {
+size_t oct_remove_recursive(octree *ot, void *object) {
   int i;
   int removed = l_remove_all_elements(ot->contents, object);
   if (removed > 0) {
